data-fetcher/requirements.txt
yfinance>=0.2.28
firebase-admin>=6.5.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.10.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
argparse
secfsdstools>=1.6.0